ISIS-2444: docs; moves nested classes out of ExceptionRecognizer so that they can be more easily referenced.
