+++ api/applib/src/main/adoc/modules/applib-svc/pages/ExceptionRecognizer.adoc
@@ -4,38 +4,13 @@
 :Notice: Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at. http://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR  CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 
 
-WARNING: TODO: this content has not yet been reviewed/updated for v2.0
-
-The `ExceptionRecognizer` SPI provides the mechanism for both the domain programmer and also for components to be able to recognize and handle certain exceptions that may be thrown by the system.
-Rather than display an obscure error to the end-user, the application can instead display a user-friendly message.
-
-For example, the JDO/DataNucleus Objectstore provides a set of recognizers to recognize and handle SQL constraint exceptions such as uniqueness violations.
-These can then be rendered back to the user as expected errors, rather than fatal stacktraces.
-
-
-
-
-== API
-
 include::system:generated:page$index/applib/services/exceprecog/ExceptionRecognizer.adoc[leveloffset=+2]
 
-TODO example migration
 
-.Deprecated Docs
-[WARNING]
-================================
+include::ExceptionRecognizer/hooks/implementation.adoc[]
 
-== SPI
+include::ExceptionRecognizer/hooks/examples_and_usage.adoc[]
 
-The SPI defined by this service is:
-
-[source,java]
-----
-include::refguide:applib-svc:example$services/exceprecog/ExceptionRecognizer.java[tags="refguide"]
-----
-<.> the main API, to attempt to recognize an exception
-
-where `Recognition` translates the recognized exception into a form that can be handled by the framework:
 
 [source,java]
 ----
